requirements.txt
langchain-core
langchain-mistralai
langchain-community
langchain-text-splitters
langgraph
faiss-cpu
streamlit
python-dotenv
